Create Post-Mortem 5-6-24
The very first game I ever worked on in this class was Table Titans, and it was supposed to be a top-down battle-bots type fighting game. The first thing I ever added was a touch interface, where I learned the simple idea of getting data from the touch of a finger
storing that data using variables, and using that data to make a functioning joystick. We then found a way to make the mech move according to the joystick with add movement input. I also learned how to use cameras, place actors and objects. and simple things
like navigating the editor. The first few weeks were spent learning the basics, and learning how to use a blueprint. Then, before we got really any progress on our main game, our first game jam started, the warren tech 50th game. We decided to make a retro-based tycoon game
simulating warren tech. In the project I learned so much about widgets, using buttons in widgets to make an action happen, and how to use widgets to add additional art effects. Another thing I learned is how to apply my hit result and variable knowledge from table titans to this game in the form of camera controls. Using some math, based on where our cursor is when we hold the left click our camera moves in certain directions. I also learned how to work with my artists and help them get their artwork in the game and look good. Some minor things I learned are using timers, bindings, and level building. Overall, I think I did pretty good work for this game, but our game came out boring, unbalanced, and looking terrible. I thought it was ok for our first ever game. We continued to work on our main projects then, where we tested how the process of the game goes, how many players we have, what abilities etc. I feel like for most of that project a lot of us were designers since none of our designers actually had any idea of what the game should be, resulting in us getting unmotivated and not a lot of work done. This project was what really showed me that if everyone doesn’t do their job the game will crash and burn quickly. Before we decided to give up on titans I got a lot done, learned about niagara system, special effects were added for death, bullets, and most abilities, I also fully added loading to our game to prevent buffering. I was sad to see our hard work go, but we had to move on. Going back a little, our next game jam of the year was upside down, and everyone had vague ideas of what to do. Our original idea was a fps game that rotates. I got weapons working with advanced inputs, and projectile collision working perfect, the only issue is our multiplayer not working, and our upsidedown mechanic also not working, resulting in our team giving up. We then split into our own groups. I decided to make a game in 1 day with just me and a designer, and I learned how to add AI movement, working stamina system using variables and inputs, and game logic. In 1 and a half days we got a game with a maze where a monster chases you until you reach the end, the player must manage their stamina, use a mirror above the maze to escape, and a jumpscare that actually scares everyone for some reason! I was pretty proud of what I could do in 1 day. The next game jam we worked on was very interesting. It was supposed to be not a game so we decided to make a “VR experience.” We REALLY struggled with this one. Everything is harder in VR, art, coding, design. Our performance was really bad and I learned some ways on how to prioritize it. I also learned how to make a simple grabbing feature for VR. Although the final product wasn’t what I envisioned, I am still pretty proud of how it feels and is very relaxing with the water sounds I added. Now, we are working on our Tower Defense game, Into The Clouds. This project is by far my favorite project I worked on and I also learned the most. I learned how to make an object come from a 2d widget using info from under the cursor. Controlling AI movement for our enemies to travel across our track, using A HUNDRED variables to store any data we need in all of our blueprints are huge in this project. I tested chaos destruction, and I added a way for us to break a part of our enemy based on our health to visualize the enemies health, I added a start screen for our game, and animated our shop. I also made some of our tower function, Like our laser tower perfectly tracking the first enemy in range. I also made our range visual and added logic for when it is to be shown and when it should not be. I also made upgrade work completely, by sending our old level 1 tower off to a location offscreen to be destroyed, then spawn in our new level 2 tower at the old ones location, using the level 2s unique upgrade blueprint thinin because it is a child of our level 1 tower. I’ve learned so many thing this year, and I know I am missing so much but these are the main things that come to mind. Probably the most important thing I learned is how to work in a team, and take charge when others don’t. I’m overall very proud of the work I’ve done in this class and can’t wait to do even more incredible things next year!